edit job

+++ database/migrations/2023_09_03_175747_expense_rates_table.php
@@ -26,7 +26,7 @@ class ExpenseRatesTable extends Migration
             $table->decimal('advertising', 8, 2); // Реклама
             $table->decimal('profit', 8, 2); // Прибыль
             $table->string('price_name');
-            $table->integer('weight');
+            $table->decimal('weight',8, 2); // вес
             $table->integer('category_id');
             $table->timestamps();
         });
